% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{setupParallel}
\alias{setupParallel}
\title{Determine the status of parallel processing. If cores are requested,
set up the cores. If cores are registered, keep using those.}
\usage{
setupParallel(parallel, silent = FALSE)
}
\arguments{
\item{parallel}{logical or integer: whether to use pre-existing parallel
workers, or how many to create and use. Default \code{TRUE}.}

\item{silent}{logical: give warnings or updates in the computation?
Default \code{FALSE}.}
}
\value{
list (ncores, cluster) ncores is the number of cores and
cluster object is a parallel::makeCluster(parallel) object or FALSE
if there is no cluster.
}
\description{
Determine the status of parallel processing. If cores are requested,
set up the cores. If cores are registered, keep using those.
}
\keyword{internal}
